Add `swift-experimental-destination` tool
