+++ Sources/swift-package-manager/main.swift
@@ -11,6 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 import Commands
+import CrossCompilationDestinationsTool
 import PackageCollectionsTool
 import PackageRegistryTool
 import TSCBasic
@@ -24,6 +25,8 @@ case "swift-package":
     SwiftPackageTool.main()
 case "swift-build":
     SwiftBuildTool.main()
+case "swift-destination":
+    SwiftDestinationTool.main()
 case "swift-test":
     SwiftTestTool.main()
 case "swift-run":
